refactor(types): extract chart types to separate file
Move type definitions from chart.tsx and chart-lazy.tsx to a dedicated chart-types.ts file to improve code organization and reduce duplication. This change maintains the same functionality while making the codebase more maintainable.
